Split package: what needs DiffEq vs what doesnt
