+++ PMFRGCore.jl/Example/differentObservables.jl
