+++ PMFRGCore.jl/Example/evaluation_susceptibility.jl
