+++ PMFRGCore.jl/Example/evaluation_thermodynamics.jl
