+++ PMFRGCore.jl/Example/evaluation_ward_identity.jl
