+++ PMFRGCore.jl/Example/generating_data.jl
