+++ PMFRGCore.jl/ext/PMFRGMPIExt/mpi/best_partition_triangle.jl
