+++ PMFRGCore.jl/ext/PMFRGMPIExt/test/MPITest/generate_data_example_mpi.jl
