+++ PMFRGCore.jl/ext/PMFRGMPIExt/test/MPITest/test_best_partition_triangle.jl
