+++ PMFRGCore.jl/src/VertexFunctions_Dense.jl
