+++ PMFRGCore.jl/src/VertexUtils.jl
