+++ PMFRGCore.jl/test/RegressionTests/PMFRG.getXBubble.common.jl
