+++ PMFRGCore.jl/test/RegressionTests/PMFRG.getXBubble.jl
