+++ PMFRGCore.jl/test/UnitTests/ExampleObservables.jl
